// x-pack/platform/plugins/shared/cases/common/bundled-types.gen.ts
export const Owner = lazySchema(() => z.enum(['cases', 'observability', 'securitySolution']));
export type Owner = z.infer<typeof Owner>;